make the scala helpers
