+++ swig/CMakeLists.txt
@@ -56,3 +56,22 @@ add_jar(
     "${CMAKE_SWIG_OUTDIR}/Tensor.java"
     "${CMAKE_SWIG_OUTDIR}/Trainer.java"
 )
+
+# TODO(joelgrus): This is probably not defensive or robust enough.
+option(INCLUDE_SCALA_HELPERS "INCLUDE_SCALA_HELPERS" ON)
+if(INCLUDE_SCALA_HELPERS)
+  # Compile the scala helpers
+  add_custom_command(
+    COMMENT "Building scala helpers"
+    OUTPUT scala_helper_compiled
+    DEPENDS dynet_swigJNI
+    COMMAND scalac -classpath ${CMAKE_CURRENT_BINARY_DIR}/dynet_swigJNI.jar
+                   -d         ${CMAKE_CURRENT_BINARY_DIR}/dynet_scala_helpers.jar
+                              ${CMAKE_CURRENT_LIST_DIR}/DynetScalaHelpers.scala
+  )
+
+  add_custom_target(scala_helper ALL DEPENDS scala_helper_compiled)
+
+endif()
+
+
